Add files via upload
Starting with the Numpy, Matplotlib and OpenCV.

What is NumPy?
NumPy is a Python library used for working with arrays.

What is Matplotlib?
Matplotlib helps with Data Visualization. It is a low-level graph plotting library.

What is OpenCV?
OpenCV is a cross-platform library using which we can develop real-time computer vision applications. It mainly focuses on image processing, video capture, and analysis including features like face detection and object detection.

~Muskan Handa:  https://www.linkedin.com/in/muskan-handa-8b087217a